move constants and keys
